requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
graphene-django==3.0.0
django-graphql-jwt==0.3.4
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
Faker==30.8.1
redis==4.6
django-redis==5.2.0
celery[redis]==5.2.7
django-ratelimit==4.0.0
gunicorn>=21.2.0
gevent>=23.0.0
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-mock==3.11.1
factory-boy==3.3.0
freezegun==1.2.2
responses==0.23.3